common: preserve submounts under the chroot point
When the chroot point is not a mount point but a simple directory, the
bind mount of that directory on itself loses any submounts already
present.

A `mount --rbind` keeps those submounts but they are not tracked in the
bookkeeping. After a successful mount, we use findmnt to add the newly
mounted path and all its submounts to the bookkeeping.